backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
google-generativeai==0.7.2
pydantic==2.9.2
fastapi-cors==0.0.6
pytest==8.3.2
httpx==0.27.0
requests==2.32.3
python-multipart==0.0.6